(GH12) renames some cards for consistency. Fixes some environment accessors
