+++ announcements/boss-will-spawn.js
