+++ announcements/contestant-leave.js
