+++ announcements/fight-concludes.js
