+++ announcements/stay.js
@@ -1,15 +1,15 @@
-const announceStay = (className, monster, { fleeRoll, player, activeContestants }) => {
+const announceStay = (className, card, { fleeRoll, player, activeContestants }) => {
 	if (fleeRoll) {
 		const assailants = activeContestants
 			.filter(contestant => contestant.monster !== player)
 			.map(contestant => contestant.monster.identityWithHp);
 
-		monster.environment.channel({
+		player.environment.channel({
 			announce:
 `${player.identityWithHp} tries to flee from ${assailants.join(' and ')}, but fails!`
 		});
 	} else {
-		monster.environment.channel({
+		player.environment.channel({
 			announce:
 `${player.identityWithHp} bravely stays in the ring.`
 		});
